Consolidate childcare_form.json
